requirements.txt
py-clob-client
requests
pyyaml
python-dotenv
web3
